requirements.txt
alembic==1.18.4
bcrypt==3.2.2
fastapi==0.136.1
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
loguru==0.7.3
passlib==1.7.4
pgvector==0.3.6
pydantic==2.13.3
pydantic_settings==2.14.0
email-validator==2.2.0
resend==2.23.0
SQLAlchemy==2.0.49
sqlmodel==0.0.38
supabase==2.29.0
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
...